# Optimize stacked recordings