This will optimize when a camera has multiple recordings in the cache to be moved. All the data going in should be the same so I don't believe there is any concern for race conflicts. 

My testing showed these differences in timings in total processing for each camera:

Current:

```
The sync took 1.553313970565796 seconds
The sync took 1.5253708362579346 seconds
The sync took 1.2776689529418945 seconds
```

This PR:

```
The async took 0.7610487937927246 seconds
The async took 0.6254909038543701 seconds
The async took 0.6794140338897705 seconds
```