@skrashevich 2023-06-14
```
def get_video_properties(url, get_duration=False):
    width = height = 0
    # Open the video stream
    video = cv2.VideoCapture(url)

    # Check if the video stream was opened successfully
    if not video.isOpened():
        logger.debug(f"Error opening video stream {url}.")
        return None

    # Get the width of frames in the video stream
    width = video.get(cv2.CAP_PROP_FRAME_WIDTH)

    # Get the height of frames in the video stream
    height = video.get(cv2.CAP_PROP_FRAME_HEIGHT)

    # Release the video stream
    video.release()

    result = {"width": width, "height": height}
    if get_duration:
        # Get the frames per second (fps) of the video stream
        fps = video.get(cv2.CAP_PROP_FPS)
        total_frames = int(video.get(cv2.CAP_PROP_FRAME_COUNT))
        duration = total_frames / fps

        result["duration"] = duration

    return result
```

and instead of 
```
                    ffprobe_cmd = [
                        "ffprobe",
                        "-v",
                        "error",
                        "-show_entries",
                        "format=duration",
                        "-of",
                        "default=noprint_wrappers=1:nokey=1",
                        f"{cache_path}",
                    ]
                    p = sp.run(ffprobe_cmd, capture_output=True)
                    if p.returncode == 0 and p.stdout.decode():
                        duration = float(p.stdout.decode().strip())
                    else:
                        duration = -1
```
put
```
                    streamInfo = get_video_properties(input.get("path"), True)

                    if streamInfo["duration"]:
                        duration = float(streamInfo["duration"])
                    else:
                        duration = -1
```